Allow setting of cylc section hooks at site level.
