+++ doc/siterc.tex
@@ -637,3 +637,24 @@ running the test battery.
 
 The minimum set of directives that must be supplied to the batch system on the
 site to initiate jobs for the tests.
+
+\subsubsection[{[[}cylc suite settings{]]}]{[cylc] $\rightarrow$ [UTC mode]}
+\label{SiteUTCMode}
+
+Allows you to set a default value for UTC mode in a suite at the site level.
+See ~\ref{UTC-mode} for details.
+
+\subsubsection[{[[}cylc suite settings{]]}]{[cylc] $\rightarrow$ [event hooks]}
+\label{SiteCylcHooks}
+
+You can define site defaults for each of the following options, details 
+of which can be found under ~\ref{SuiteEventHandling}:
+
+\subparagraph{[cylc] $\rightarrow$ [event hooks] $\rightarrow$ startup handler}
+
+\subparagraph{[cylc] $\rightarrow$ [event hooks] $\rightarrow$ shutdown handler}
+
+\subparagraph{[cylc] $\rightarrow$ [event hooks] $\rightarrow$ timeout handler}
+
+\subparagraph{[cylc] $\rightarrow$ [event hooks] $\rightarrow$ timeout}
+
