+++ lib/cylc/cfgspec/suite.py
@@ -35,9 +35,12 @@ from isodatetime.data import Calendar, TimePoint
 from isodatetime.parsers import TimePointParser, DurationParser
 from cylc.cycling.integer import REC_INTERVAL as REC_INTEGER_INTERVAL
 
+from cylc.cfgspec.utils import coerce_interval
+from cylc.cfgspec.utils import coerce_interval_list
+from cylc.cfgspec.globalcfg import GLOBAL_CFG
+
 "Define all legal items and values for cylc suite definition files."
 
-interval_parser = DurationParser()
 
 def _coerce_cycleinterval( value, keys, args ):
     """Coerce value to a cycle interval."""
@@ -153,51 +156,6 @@ def _coerce_final_cycletime( value, keys, args ):
     return value
 
 
-def coerce_interval(value, keys, args, back_comp_unit_factor=1,
-                    check_syntax_version=True):
-    """Coerce an ISO 8601 interval (or number: back-comp) into seconds."""
-    value = _strip_and_unquote( keys, value )
-    try:
-        backwards_compat_value = float(value) * back_comp_unit_factor
-    except (TypeError, ValueError):
-        pass
-    else:
-        if check_syntax_version:
-            set_syntax_version(VERSION_PREV,
-                               "integer interval: %s" % itemstr(
-                                   keys[:-1], keys[-1], value))
-        return backwards_compat_value
-    try:
-        interval = interval_parser.parse(value)
-    except ValueError:
-        raise IllegalValueError("ISO 8601 interval", keys, value)
-    if check_syntax_version:
-        try:
-            set_syntax_version(VERSION_NEW,
-                               "ISO 8601 interval: %s" % itemstr(
-                                   keys[:-1], keys[-1], value))
-        except SyntaxVersionError as exc:
-            raise Exception(str(exc))
-    days, seconds = interval.get_days_and_seconds()
-    seconds += days * Calendar.default().SECONDS_IN_DAY
-    return seconds
-
-
-def coerce_interval_list(value, keys, args, back_comp_unit_factor=1,
-                         check_syntax_version=True):
-    """Coerce a list of intervals (or numbers: back-comp) into seconds."""
-    values_list = _strip_and_unquote_list( keys, value )
-    type_converter = (
-        lambda v: coerce_interval(
-            v, keys, args,
-            back_comp_unit_factor=back_comp_unit_factor,
-            check_syntax_version=check_syntax_version,
-        )
-    )
-    seconds_list = _expand_list( values_list, keys, type_converter, True )
-    return seconds_list
-
-
 coercers['cycletime'] = _coerce_cycletime
 coercers['cycletime_format'] = _coerce_cycletime_format
 coercers['cycletime_time_zone'] = _coerce_cycletime_time_zone
@@ -218,7 +176,7 @@ SPEC = {
     'description'                             : vdr( vtype='string', default="" ),
     'URL'                                     : vdr( vtype='string', default="" ),
     'cylc' : {
-        'UTC mode'                            : vdr( vtype='boolean', default=False),
+        'UTC mode'                            : vdr( vtype='boolean', default=GLOBAL_CFG.get( ['cylc','UTC mode'] )),
         'cycle point format'                  : vdr( vtype='cycletime_format', default=None),
         'cycle point num expanded year digits': vdr( vtype='integer', default=0),
         'cycle point time zone'               : vdr( vtype='cycletime_time_zone', default=None),
@@ -230,10 +188,10 @@ SPEC = {
             '__MANY__'                        : vdr( vtype='string' ),
             },
         'event hooks' : {
-            'startup handler'                 : vdr( vtype='string_list', default=[] ),
-            'timeout handler'                 : vdr( vtype='string_list', default=[] ),
-            'shutdown handler'                : vdr( vtype='string_list', default=[] ),
-            'timeout'                         : vdr( vtype='interval_minutes'  ),
+            'startup handler'                 : vdr( vtype='string_list', default=GLOBAL_CFG.get( ['cylc','event hooks', 'startup handler'] ) ),
+            'timeout handler'                 : vdr( vtype='string_list', default=GLOBAL_CFG.get( ['cylc','event hooks', 'timeout handler'] ) ),
+            'shutdown handler'                : vdr( vtype='string_list', default=GLOBAL_CFG.get( ['cylc','event hooks', 'shutdown handler'] ) ),
+            'timeout'                         : vdr( vtype='interval_minutes', default=GLOBAL_CFG.get( ['cylc','event hooks', 'timeout'] ) ),
             'reset timer'                     : vdr( vtype='boolean', default=True ),
             'abort if startup handler fails'  : vdr( vtype='boolean', default=False ),
             'abort if shutdown handler fails' : vdr( vtype='boolean', default=False ),
